Card 별 refactoring + api 업데이트
